refactor: Move Database Methods to API folder
