+++ cypress/e2e/common/common.js
@@ -30,7 +30,6 @@ import { guid } from "../../helpers/commentText";
 import { hasToggle } from "../../helpers/hasToggle";
 import { searchPersonResults } from "../../support/searchPersonResults";
 import { searchPropertyResults } from "../../support/searchPropertyResults";
-import DynamoDb from "./DynamoDb";
 
 import { generateTenure } from "../../../api/models/requests/addTenureModel";
 import {
@@ -42,6 +41,9 @@ import {
 } from "../../../api/models/requests/createAssetModel";
 import { person } from "../../../api/models/requests/createPersonModel";
 import { patch } from "../../../api/models/requests/patchModel";
+import { addTestRecordToDatabase, seedDatabaseWithTenure } from "../../helpers/DbHelpers";
+
+// note - this whole file will need to be refactored in stages
 
 const envConfig = require("../../../environment-config");
 const activityHistory = new ActivityHistoryPageObjects();
@@ -1044,22 +1046,6 @@ export const getAssetViewUrlByGuid = (assetGuid) => {
   return `${envConfig.baseUrl}/property/${assetGuid}`;
 };
 
-// This methods adds a temporary database record to the provided database table, and adds an entry to recordsToDelete.json
-export const addTestRecordToDatabase = (dbTableName, testDbRecord) => {
-  cy.log("Seeding database").then(async () => {
-    cy.log(
-      `Adding test record to database table ${dbTableName} and creating a record of it in recordsToDelete.json file`
-    );
-    return new Cypress.Promise((resolve) => {
-      DynamoDb.createRecord(dbTableName, testDbRecord).then(() => {
-        resolve();
-      });
-    }).then(() => {
-      cy.log("Database seeded!");
-    });
-  });
-};
-
 // Database seed methods
 
 Given("I seeded the database", () => {
@@ -1102,52 +1088,6 @@ Given("I seeded the database", () => {
   addTestRecordToDatabase("Persons", personModel2);
 });
 
-const seedDatabaseWithTenure = (isActive) => {
-  const assetModel = generateAsset();
-  const personModel1 = person();
-  const personModel2 = person();
-  const tenureModel = generateTenure(
-    {},
-    assetModel,
-    [
-      personModel1,
-      { isResponsible: true, personTenureType: "Tenant", ...personModel2 },
-    ],
-    undefined,
-    "1990-10-13",
-    isActive ? "3050-10-13" : "1998-10-13"
-  );
-
-  // Add tenure to persons
-  const personTenure = {
-    id: tenureModel.id,
-    startDate: tenureModel.startOfTenureDate,
-    endDate: tenureModel.endOfTenureDate,
-    assetFullAddress: tenureModel.tenuredAsset.fullAddress,
-    assetId: tenureModel.tenuredAsset.id,
-    uprn: tenureModel.tenuredAsset.uprn,
-    isActive: isActive,
-    type: tenureModel.tenureType.description,
-    propertyReference: tenureModel.tenuredAsset.propertyReference,
-  };
-
-  personModel1.tenures.push(personTenure);
-  personModel2.tenures.push(personTenure);
-
-  // Add tenure to asset
-  assetModel.tenure = {
-    endOfTenureDate: tenureModel.endOfTenureDate,
-    id: tenureModel.id,
-    paymentReference: tenureModel.paymentReference,
-    startOfTenureDate: tenureModel.startOfTenureDate,
-    type: tenureModel.tenureType.description,
-  };
-
-  addTestRecordToDatabase("Assets", assetModel);
-  addTestRecordToDatabase("TenureInformation", tenureModel);
-  addTestRecordToDatabase("Persons", personModel1);
-  addTestRecordToDatabase("Persons", personModel2);
-};
 
 Given("I seeded the database with a tenure", () => {
   seedDatabaseWithTenure(true);
@@ -1223,42 +1163,4 @@ Given(
 Given("I seeded the database with an asset", () => {
   const testAsset = generateAsset();
   addTestRecordToDatabase("Assets", testAsset);
-});
-
-// Database data teardown
-
-beforeEach(() => {
-  const filename = "cypress/fixtures/recordsToDelete.json";
-  cy.readFile(filename).then((recordsToDelete) => {
-    if (recordsToDelete.length) {
-      return new Cypress.Promise((resolve) => {
-        Promise.all(
-          recordsToDelete.map((record) => dynamoDb.deleteRecord(record))
-        ).then(() => {
-          resolve();
-        });
-      }).then(() => {
-        cy.writeFile(filename, []);
-        cy.log("Test database records cleared!");
-      });
-    }
-  });
-});
-
-afterEach(() => {
-  const filename = "cypress/fixtures/recordsToDelete.json";
-  cy.readFile(filename).then((recordsToDelete) => {
-    if (recordsToDelete.length) {
-      return new Cypress.Promise((resolve) => {
-        Promise.all(
-          recordsToDelete.map((record) => dynamoDb.deleteRecord(record))
-        ).then(() => {
-          resolve();
-        });
-      }).then(() => {
-        cy.writeFile(filename, []);
-        cy.log("Test database records cleared!");
-      });
-    }
-  });
 });
